requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
